Enum hex2d::Direction

pub enum Direction {
    YZ,
    XZ,
    XY,
    ZY,
    ZX,
    YX,
}

Direction on a hexagonal map

See Coordinate for graph with directions.

Naming convention: increasing coordinate for a given direction is first decreasing is second. The missing coordinate is unaffected by a move in a given direction.

Variants

YZ

+Y -Z

XZ

-Z +X

XY

+X -Y

ZY

-Y +Z

ZX

+Z -X

YX

-X +Y

Methods

impl Direction

fn all() -> &'static [Direction; 6]

Static array of all directions

use hex2d::Direction;

assert_eq!(Direction::all().len(), 6);

fn arc(&self, steps: u8) -> Vec<Direction>

Return a vector of an arc including Directions steps away from the original Direction both sides from left to right.

use hex2d::{Direction};
use hex2d::Angle::*;

for &d in Direction::all() {
    assert_eq!(d.arc(0), vec!(d));
    assert_eq!(d.arc(1), vec!(d + Left, d, d + Right));
    assert_eq!(d.arc(2), vec!(d + LeftBack, d + Left, d, d + Right, d + RightBack));
    assert_eq!(d.arc(3), vec!(d + LeftBack, d + Left, d, d + Right, d + RightBack, d + Back));
}

fn from_int<I: Integer>(i: I) -> Direction

Create Direction from integer in [0, 6) range

This should probably be internal

fn to_int<I: Integer>(&self) -> I

Convert to integer in [0, 6) range

This should probably be internal