Struct hex2d::Coordinate

pub struct Coordinate<I: Integer = i32> {
    pub x: I,
    pub y: I,
}

Coordinate on 2d hexagonal grid

Fields

x

x coordinate

y

y coordinate

Methods

impl<I: Integer> Coordinate<I>

fn new(x: I, y: I) -> Coordinate<I>

Create new Coord from x and y

fn scale(&self, s: I) -> Coordinate<I>

Scale coordinate by a factor s

fn from_round(x: f32, y: f32) -> Option<Coordinate<I>>

Round x, y float to nearest hex coordinates

fn for_each_in_line_to<F>(&self, dest: Coordinate<I>, f: F) where F: FnMut(Coordinate<I>), &'a I: Add<&'a I, Output=I>

Execute f for each coordinate in straight line from self to dest

fn line_to(&self, dest: Coordinate<I>) -> Vec<Coordinate<I>> where &'a I: Add<&'a I, Output=I>

Construct a straight line to a dest

fn z(&self) -> I

Z coordinate

fn direction_from_center_cw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal

Returns: None if is center

fn direction_from_center_ccw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

fn direction_to_cw(&self, coor: Coordinate<I>) -> Option<Direction>

Direction from self to coord

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal.

Returns: None if is center

fn direction_to_ccw(&self, coor: Coordinate<I>) -> Option<Direction>

Direction from self to coor

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

fn neighbors(&self) -> [Coordinate<I>; 6]

Array with all the neighbors of a coordinate

fn distance(&self, c: Coordinate<I>) -> I

Distance between two Coordinates

fn range(&self, r: I) -> Vec<Coordinate<I>> where &'a I: Add<&'a I, Output=I>

All coordinates in radius r

fn for_each_in_range<F>(&self, r: I, f: F) where F: FnMut(Coordinate<I>), &'a I: Add<&'a I, Output=I>

Execute f for all coordinates in radius r

fn ring(&self, r: i32, s: Spin) -> Vec<Coordinate<I>>

A ring of radius r, starting in a corner in a given Direction

Example: Elements in order for Ring of radius 2, Direction ZX, CCW

             8
           9   7
        10   .   6
           .   .
        11   x   5
           .   .
         0   .   4
           1   3
             2

fn for_each_in_ring<F>(&self, r: i32, s: Spin, f: F) where F: FnMut(Coordinate<I>)

Call f for each coordinate in a ring

See ring for a ring description.

fn to_pixel_float(&self, spacing: Spacing) -> (f32, f32)

Convert to pixel coordinates using spacing, where the parameter means the edge length of a hexagon.

This function is meant for graphical user interfaces where resolution is big enough that floating point calculation make sense.

fn to_pixel_integer(&self, spacing: IntegerSpacing<I>) -> (I, I)

Convert to integer pixel coordinates using spacing, where the parameters mean the width and height multiplications

fn from_pixel_integer(spacing: IntegerSpacing<I>, v: (I, I)) -> (Coordinate<I>, (I, I))

Convert integer pixel coordinates v using spacing to nearest coordinate that has both integer pixel coordinates lower or equal to v. Also return offset (in integer pixels) from that coordinate.

fn rotate_around_zero(&self, a: Angle) -> Coordinate<I>

Rotate self around a point (0, 0, 0) using angle of rotation a

fn rotate_around(&self, center: Coordinate<I>, a: Angle) -> Coordinate<I>

Rotate self around a center using angle of rotation a