Struct hex2d::Coordinate

pub struct Coordinate<I: Integer> {
    pub x: I,
    pub y: I,
}

Coordinate on 2d hexagonal grid

Fields
x: I

x coordinate

y: I

y coordinate

Methods

impl<I: Integer> Coordinate<I>

pub fn new(x: I, y: I) -> Coordinate<I>

Create new Coord from x and y

pub fn from_round(x: f32, y: f32) -> Coordinate<I>

Old name for nearest

pub fn nearest(x: f32, y: f32) -> Coordinate<I>

Round x, y float to nearest hex coordinates

pub fn from_round_lossy(x: f32, y: f32) -> Option<Coordinate<I>>

Old name for nearest_lossy

pub fn nearest_lossy(x: f32, y: f32) -> Option<Coordinate<I>>

Round x, y float to nearest hex coordinates

Return None, if exactly on the border of two hex coordinates

pub fn from_pixel_integer(
spacing: IntegerSpacing<I>,
v: (I, I)
) -> (Coordinate<I>, (I, I))

Old name for nearest_with_offset

pub fn from_pixel(x: f32, y: f32, spacing: Spacing) -> Coordinate<I>

Find the hex containing a pixel. The origin of the pixel coordinates is the center of the hex at (0,0) in hex coordinates.

pub fn nearest_with_offset(
spacing: IntegerSpacing<I>,
v: (I, I)
) -> (Coordinate<I>, (I, I))

Convert integer pixel coordinates v using spacing to nearest coordinate that has both integer pixel coordinates lower or equal to v. Also return offset (in integer pixels) from that coordinate.

Useful for ASCII visualization.

pub fn to_pixel_float(&self, spacing: Spacing) -> (f32, f32)

Old name for to_pixel

pub fn to_pixel(&self, spacing: Spacing) -> (f32, f32)

Convert to pixel coordinates using spacing, where the parameter means the edge length of a hexagon.

This function is meant for graphical user interfaces where resolution is big enough that floating point calculation make sense.

pub fn to_pixel_integer(&self, spacing: IntegerSpacing<I>) -> (I, I)

Convert to integer pixel coordinates using spacing, where the parameters mean the width and height multiplications

pub fn scale(&self, s: I) -> Coordinate<I>

Scale coordinate by a factor s

pub fn neighbors(&self) -> [Coordinate<I>; 6]

Array with all the neighbors of a coordinate

pub fn rotate_around_zero(&self, a: Angle) -> Coordinate<I>

Rotate self around a point (0, 0, 0) using angle of rotation a

pub fn rotate_around(&self, center: Coordinate<I>, a: Angle) -> Coordinate<I>

Rotate self around a center using angle of rotation a

pub fn for_each_in_line_to<F>(&self, dest: Coordinate<I>, f: F) where
F: FnMut(Coordinate<I>),
&'a I: Add<&'a I, Output = I>,

Execute f for each coordinate in straight line from self to dest

pub fn for_each_in_line_to_lossy<F>(&self, dest: Coordinate<I>, f: F) where
F: FnMut(Coordinate<I>),
&'a I: Add<&'a I, Output = I>,

Execute f for each coordinate in straight line from self to dest

Skip points on the border of two tiles

pub fn for_each_in_line_to_with_edge_detection<F>(
&self,
dest: Coordinate<I>,
f: F
) where
F: FnMut((Coordinate<I>, Coordinate<I>)),
&'a I: Add<&'a I, Output = I>,

Execute f for pairs of coordinates in straight line from self to dest

On edge condition the pair contains different members, otherwise it's the same.

pub fn line_to(&self, dest: Coordinate<I>) -> Vec<Coordinate<I>> where
&'a I: Add<&'a I, Output = I>,

Construct a straight line to a dest

pub fn line_to_lossy(&self, dest: Coordinate<I>) -> Vec<Coordinate<I>> where
&'a I: Add<&'a I, Output = I>,

Construct a straight line to a dest

Skip points on the border of two tiles

pub fn line_to_with_edge_detection(
&self,
dest: Coordinate<I>
) -> Vec<(Coordinate<I>, Coordinate<I>)> where
&'a I: Add<&'a I, Output = I>,

Construct a straight line to a dest

pub fn z(&self) -> I

Z coordinate

pub fn direction_from_center_cw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal

Returns: None if is center

use hex2d::{Direction, Coordinate};
use hex2d::{Left, Right};

let center = Coordinate::new(0, 0);

assert_eq!(center.direction_from_center_cw(), None);

for &d in Direction::all() {
    assert_eq!((center + d).direction_from_center_cw(), Some(d));
    assert_eq!((center + d + (d + Left)).direction_from_center_cw(), Some(d));
    assert_eq!((center + d + (d + Right)).direction_from_center_cw(), Some(d + Right));
}

pub fn directions_from_center(&self) -> Vec<Direction>

Directions that lead from center to a given point.

Returns an array of on eor two dirs

pub fn direction_from_center_ccw(&self) -> Option<Direction>

Direction from center (0, 0) to coordinate

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

use hex2d::{Direction, Coordinate};
use hex2d::{Left, Right};

let center = Coordinate::new(0, 0);

assert_eq!(center.direction_from_center_ccw(), None);

for &d in Direction::all() {
    assert_eq!((center + d).direction_from_center_ccw(), Some(d));
    assert_eq!((center + d + (d + Left)).direction_from_center_ccw(), Some(d + Left));
    assert_eq!((center + d + (d + Right)).direction_from_center_ccw(), Some(d));
}

pub fn directions_to(&self, coord: Coordinate<I>) -> Vec<Direction>

Directions from self to coord

pub fn direction_to_cw(&self, coor: Coordinate<I>) -> Option<Direction>

Direction from self to coord

In case of diagonals (edge of two major directions), prefers direction that is clockwise from the diagonal.

Returns: None if is center

pub fn direction_to_ccw(&self, coor: Coordinate<I>) -> Option<Direction>

Direction from self to coor

In case of diagonals (edge of two major directions), prefers direction that is counter-clockwise from the diagonal.

Returns: None if is center

pub fn distance(&self, c: Coordinate<I>) -> I

Distance between two Coordinates

pub fn range(&self, r: I) -> Vec<Coordinate<I>> where
&'a I: Add<&'a I, Output = I>,

All coordinates in radius r

pub fn for_each_in_range<F>(&self, r: I, f: F) where
F: FnMut(Coordinate<I>),
&'a I: Add<&'a I, Output = I>,

Execute f for all coordinates in radius r

pub fn ring(&self, r: i32, s: Spin) -> Vec<Coordinate<I>>

A ring of radius r, starting in a corner in a given Direction

Example: Elements in order for Ring of radius 2, Direction ZX, CCW

             8
           9   7
        10   .   6
           .   .
        11   x   5
           .   .
         0   .   4
           1   3
             2

use hex2d::{Coordinate, Spin, XY};

let center = Coordinate::new(5, -1);

for &c in &center.neighbors() {
    for &ring_c in &c.ring(5, Spin::CCW(XY)) {
        assert_eq!(c.distance(ring_c), 5);
    }
}

pub fn for_each_in_ring<F>(&self, r: i32, s: Spin, f: F) where
F: FnMut(Coordinate<I>),

Call f for each coordinate in a ring

See ring for a ring description.